# docs/openapi.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stellar Bounty Board API",
    "version": "1.0.0",
    "description": "REST API for the Stellar Bounty Board — a platform for posting, reserving, submitting, and releasing on-chain bounties backed by Stellar tokens.\n\n**Bounty lifecycle:** `open` → `reserved` → `submitted` → `released`\n\nMaintainers may `cancel` an `open` bounty before reservation, or `refund` an `open` or `reserved` bounty after the deadline. Bounties whose deadline passes are automatically transitioned to `expired`."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
  "components": {
    "schemas": {
...
      "BountyAuditLogListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BountyAuditLogRecord"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/BountyAuditLogPagination"
          }
        },
        "required": [
          "data",
          "pagination"
        ]
      },
...
      "ReserveBountyRequest": {
        "type": "object",
        "properties": {
          "contributor": {
            "type": "string",
            "description": "Stellar public key of the contributor reserving the bounty.",
            "example": "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF"
          },
          "expectedVersion": {
            "type": "integer",
            "description": "Optional version number for race condition prevention. If provided, must match current bounty version.",
            "example": 1
          }
        },
        "required": [
          "contributor"
        ]
      },
      "SubmitBountyRequest": {
        "type": "object",
        "properties": {
          "contributor": {
            "type": "string",
            "description": "Must match the contributor who reserved the bounty.",
            "example": "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF"
          },
          "notes": {
            "type": "string",
            "maxLength": 240,
            "description": "Optional notes (max 240 chars).",
            "example": "Fixed by updating the redirect handler."
          }
        },
        "required": [
          "contributor"
        ]
      },
      "MaintainerActionRequest": {
        "type": "object",
        "properties": {
          "maintainer": {
            "type": "string",
            "description": "Must match the maintainer address on the bounty.",
            "example": "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF"
          },
          "transactionHash": {
            "type": "string",
            "pattern": "^[0-9a-fA-F]{64}$",
            "description": "Optional transaction hash for the release/refund action (64-char hex).",
            "example": "0000000000000000000000000000000000000000000000000000000000000000"
          }
        },
        "required": [
          "maintainer"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Bounty not found."
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/bounties": {
      "get": {
        "tags": [
          "Bounties"
        ],
        "summary": "List all bounties",
        "description": "Returns every bounty record sorted by creation date (newest first). Bounties whose deadline has passed are automatically transitioned to `expired` before the list is returned.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Case-insensitive substring filter applied to title, summary, and labels."
            },
            "required": false,
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Exact Stellar public key filter applied to the bounty contributor."
            },
            "required": false,
            "name": "contributor",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter bounties with deadline before this ISO 8601 date string.",
              "example": "2026-07-28T19:20:12.196Z"
            },
            "required": false,
            "name": "deadlineBefore",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter bounties with deadline after this ISO 8601 date string.",
              "example": "2026-06-28T19:20:12.200Z"
            },
            "required": false,
            "name": "deadlineAfter",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (starts at 1, default 1)."
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "description": "Number of items per page (max 100, default 20)."
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of all bounty records.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BountyRecord"
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters (e.g., invalid date string).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Bounties"
        ],
        "summary": "Create a bounty",
        "description": "Creates a new open bounty and persists it. Rate-limited to **5 requests per IP per minute**. The `deadlineAt` timestamp is computed as `now + deadlineDays * 86400`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBountyRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bounty created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/BountyRecord"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation failed — see `error` field for details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/bounties/{id}/audit-logs": {
      "get": {
        "tags": [
          "Bounties"
        ],
        "summary": "List audit logs for one bounty",
        "description": "Returns ordered status transition history for a bounty. Use `limit` (1-100, default 20) and `offset` (default 0) for pagination.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "BNT-0001"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "description": "Maximum number of audit entries to return (1-100).",
              "example": 20
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Zero-based offset into the ordered audit history.",
              "example": 0
            },
            "required": false,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Audit log page for the requested bounty.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BountyAuditLogListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bounty not found, bounty id invalid, or pagination query invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/bounties/{id}/reserve": {
      "post": {
        "tags": [
          "Bounties"
        ],
        "summary": "Reserve a bounty",
        "description": "Assigns a contributor to an `open` bounty, transitioning its status to `reserved`. Only one contributor can hold a reservation at a time. Rate-limited to **5 requests per IP per minute**.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "BNT-0001"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveBountyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bounty successfully reserved.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/BountyRecord"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bounty not found, not open, or validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/bounties/{id}/submit": {
      "post": {
        "tags": [
          "Bounties"
        ],
        "summary": "Submit work for a bounty",
        "description": "Transitions a `reserved` bounty to `submitted` by attaching a submission URL. The `contributor` field must exactly match the address that reserved the bounty. Rate-limited to **5 requests per IP per minute**.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "BNT-0001"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitBountyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Work submitted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/BountyRecord"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bounty not found, not reserved, contributor mismatch, or validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/bounties/{id}/release": {
      "post": {
        "tags": [
          "Bounties"
        ],
        "summary": "Release payment for a bounty",
        "description": "Transitions a `submitted` bounty to `released`, indicating payment has been sent. Only the maintainer address recorded on the bounty may call this endpoint. Rate-limited to **5 requests per IP per minute**.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "BNT-0001"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintainerActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment released.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/BountyRecord"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bounty not found, not submitted, maintainer mismatch, or validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}